Create multiple js files
