+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<link rel="stylesheet" href="style.css">
+	<link rel="stylesheet" href="index.css">
     <title>Awesome Books</title>
 </head>
 <body>
@@ -13,11 +13,9 @@
 			<div class="nav-b">
 				<p>Awesome Books</p>
 			</div>
-			<a href="#" class="nav-lnk" id="nav-lst">List</a>
-			<span class="vertical-line"></span>
-			<a href="#" class="nav-lnk" id="nav-plus">Add new</a>
-			<span class="vertical-line"></span>
-			<a href="#" class="nav-lnk" id="nav-contact-us">Contact us</a>
+			<button class="nav-lnk" id="nav-lst">List</button>
+<button class="nav-lnk" id="nav-plus">Add new</button>
+<button class="nav-lnk" id="nav-contact-us">Contact us</button>
 		</div>
 	</header>
 		
@@ -73,6 +71,6 @@
 	<footer>
 		<h2>Copyrights@AlyAtalla </h2>
 	</footer>
-	<script src="./script.js"></script>
+	<script src="./index.js" type="module"></script>
 </body>
 </html>
