+++ package.json
@@ -1,31 +1,12 @@
 {
-  "name": "awesome-books-plain-javascript-with-objects",
+  "name": "modules",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/omar25ahmed/Awesome-books-plain-JavaScript-with-objects.git"
-  },
   "keywords": [],
   "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/omar25ahmed/Awesome-books-plain-JavaScript-with-objects/issues"
-  },
-  "homepage": "https://github.com/omar25ahmed/Awesome-books-plain-JavaScript-with-objects#readme",
-  "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.25.4",
-    "hint": "^6.1.9",
-    "stylelint": "^13.13.1",
-    "stylelint-config-standard": "^21.0.0",
-    "stylelint-csstree-validator": "^1.9.0",
-    "stylelint-scss": "^3.21.0"
-  }
+  "license": "ISC"
 }
