--- script.js
@@ -1,123 +1,0 @@
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-
-class Store {
-  static getBooks() {
-    let books;
-    if (localStorage.getItem('books') === null) {
-      books = [];
-    } else {
-      books = JSON.parse(localStorage.getItem('books'));
-    }
-    return books;
-  }
-
-  static addBook(book) {
-    const books = Store.getBooks();
-    books.push(book);
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static removeBook(author) {
-    const books = Store.getBooks();
-    const filteredBooks = books.filter(book => book.author !== author);
-    localStorage.setItem('books', JSON.stringify(filteredBooks));
-  }
-}
-
-class ShowBooks {
-  static displayBooks() {
-    const books = Store.getBooks();
-    books.forEach(book => ShowBooks.addBookToList(book));
-  }
-
-  static addBookToList(book) {
-    const list = document.querySelector('#list');
-    const row = document.createElement('tr');
-
-    row.innerHTML = `
-        <td>${book.title} by </td>
-        <td>${book.author}</td>
-        <td><button class="delete">Remove book</button></td>
-      `;
-
-    list.appendChild(row);
-  }
-
-  static deleteBook(el) {
-    if (el.classList.contains('delete')) {
-      el.parentElement.parentElement.remove();
-    }
-  }
-
-  static clearFields() {
-    document.querySelector('#title').value = '';
-    document.querySelector('#author').value = '';
-  }
-}
-
-document.addEventListener('DOMContentLoaded', ShowBooks.displayBooks);
-
-const showBookList = document.getElementById('nav-lst');
-const showAddBook = document.getElementById('nav-plus');
-const showContact = document.getElementById('nav-contact-us');
-
-function displayList() {
-  document.getElementById('book-list').style.display = 'flex';
-  document.getElementById('add-new').style.display = 'none';
-  document.getElementById('contact').style.display = 'none';
-}
-
-function displayPlus() {
-  document.getElementById('book-list').style.display = 'none';
-  document.getElementById('add-new').style.display = 'flex';
-  document.getElementById('contact').style.display = 'none';
-}
-
-function showContactPage() {
-  document.getElementById('book-list').style.display = 'none';
-  document.getElementById('add-new').style.display = 'none';
-  document.getElementById('contact').style.display = 'flex';
-}
-
-window.addEventListener('load', () => {
-  displayList();
-});
-
-showBookList.addEventListener('click', () => {
-  displayList();
-});
-
-showAddBook.addEventListener('click', () => {
-  displayPlus();
-});
-
-showContact.addEventListener('click', () => {
-  showContactPage();
-});
-
-document.querySelector('#form').addEventListener('submit', (e) => {
-  e.preventDefault();
-
-  const title = document.querySelector('#title').value;
-  const author = document.querySelector('#author').value;
-
-  const book = new Book(title, author);
-
-  ShowBooks.addBookToList(book);
-  Store.addBook(book);
-  ShowBooks.clearFields();
-});
-
-document.querySelector('#list').addEventListener('click', (e) => {
-  ShowBooks.deleteBook(e.target);
-  const author = e.target.parentElement.previousElementSibling.textContent;
-  Store.removeBook(author);
-});
-
-const date = new Date();
-document.getElementById('date').innerHTML = date;
