--- style.css
@@ -1,98 +1,0 @@
-body {
-  margin: 0;
-  font-family: "Times New Roman", Times, serif;
-}
-
-.head-top {
-  margin: 15px;
-  padding: 0;
-}
-
-.navbar {
-  margin: 0;
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  border: 5px solid rgb(0, 0, 0);
-  background-color: white;
-}
-
-.nav-lnk {
-  padding: 5px 10px;
-}
-.vertical-line {
-  border-left: 1px solid black;
-  height: 17px;
-  margin: 0 10px;
-}
-
-.nav-b {
-  padding-right: 400px;
-}
-
-a {
-  text-decoration: none;
-  color: rgb(0, 0, 0);
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-table {
-  border: 4px solid rgb(0, 0, 0);
-}
-
-.delete {
-  margin-left: 150px;
-  background-color: #fff;
-}
-
-hr {
-  width: 100px;
-  margin: 35px 0;
-}
-
-#form {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.form-group {
-  margin: 10px 0;
-}
-
-.submit {
-  align-self: flex-end;
-}
-
-#book-list {
-  display: flex;
-  justify-content: center;
-  align-self: center;
-}
-
-#add-new {
-  display: flex;
-  justify-content: center;
-  align-self: center;
-}
-
-#contact-us {
-  display: flex;
-  justify-content: center;
-  align-self: center;
-}
-
-footer {
-  position: absolute;
-  bottom: 0;
-  width: 96%;
-  margin: 25px;
-  border: 5px solid rgb(0, 0, 0);
-}
